open-telemetry/tracing-otel-pipeline/service1/requirements.txt
Flask
uwsgi
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp